Restrict `subst_term` to `x.f` where `f` is a model field
In the QCheck-STM framework, the code generated to test whether clauses
hold will be run with only the model state in scope, not the actual
non-model values. `subst_term` will ensure that the substitution is
indeed used in such a case instead of generating non-compiling code.
This adds a new error case to report those impossible substitutions.